requirements.txt
spoon-ai-sdk
fitz
numpy
faiss-cpu
sentence-transformers
streamlit
gradio